% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abalone.i}
\alias{abalone.i}
\title{abalone.i data example}
\format{
An object of class \code{data.frame} (inherits from \code{symbolic_tbl}) with 24 rows and 7 columns.
}
\source{
Adapted from \code{MAINT.Data::AbaloneIdt}; the underlying
  Abalone data are from the UCI Machine Learning Repository.
}
\usage{
data(abalone.i)
}
\description{
abalone.i interval data example.
}
\examples{
data(abalone.i)
ggInterval_indexplot(abalone.i, aes(x = Length))
}
\keyword{datasets}
