% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blood.i}
\alias{blood.i}
\title{blood.i data example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}, \code{symbolic_tbl}) with 14 rows and 3 columns.
}
\source{
Adapted from \code{HistDAWass::BLOOD}; see
  \url{https://CRAN.R-project.org/package=HistDAWass}.
}
\usage{
data(blood.i)
}
\description{
blood.i interval data example.
}
\examples{
data(blood.i)
ggInterval_MMplot(blood.i, aes(x = Hematocrit))
}
\references{
Billard, Lynne and Diday, Edwin (2006).
\emph{Symbolic Data Analysis: Conceptual Statistics and Data Mining}.
Chichester, UK: John Wiley and Sons.
}
\keyword{datasets}
