% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cardiological2}
\alias{Cardiological2}
\title{Cardiological data example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 15 rows and 3 columns.
}
\source{
Adapted from \code{RSDA::Cardiological2}; see
  \url{https://CRAN.R-project.org/package=RSDA}.
}
\usage{
data(Cardiological2)
}
\description{
Cardiological interval data example.
}
\examples{
data(Cardiological2)
ggInterval_indexplot(Cardiological2, aes(x = Syst))
}
\references{
Billard, Lynne and Diday, Edwin (2006).
\emph{Symbolic Data Analysis: Conceptual Statistics and Data Mining}.
Chichester, UK: John Wiley and Sons.
}
\keyword{datasets}
