% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cov}
\alias{cov}
\alias{cov.default}
\alias{cov.symbolic_tbl}
\alias{cov.symbolic_interval}
\title{Generic function for the covariance}
\usage{
cov(x, ...)

\method{cov}{default}(
  x,
  y = NULL,
  use = "everything",
  method = c("pearson", "kendall", "spearman"),
  ...
)

\method{cov}{symbolic_tbl}(x, ...)

\method{cov}{symbolic_interval}(x, y = NULL, method = c("centers", "B", "BD", "BG"), na.rm = FALSE, ...)
}
\arguments{
\item{x}{First symbolic variables.}

\item{...}{As in R cov function.}

\item{y}{Second symbolic variables.}

\item{use}{an optional character string giving a method for computing
covariances in the presence of missing values. This must be (an abbreviation of)
 one of the strings 'everything', 'all.obs', 'complete.obs', 'na.or.complete',
 or 'pairwise.complete.obs'.}

\item{method}{The method to use.}

\item{na.rm}{As in R cov function.}
}
\value{
Return a real number.
}
\description{
This function compute the symbolic covariance.
}
\details{
Supported interval-valued methods are:
\itemize{
  \item \code{"centers"}: covariance of interval centers.
  \item \code{"B"}: Billard covariance.
  \item \code{"BD"}: Billard-Diday covariance.
  \item \code{"BG"}: Bertrand-Goupil covariance.
}
}
\references{
Bertrand, Patrice and Goupil, Francoise (2000). Descriptive Statistics for
Symbolic Data. In Hans-Hermann Bock and Edwin Diday (eds.),
\emph{Analysis of Symbolic Data}, pp. 106--124. Berlin and Heidelberg:
Springer.

Billard, Lynne and Diday, Edwin (2006). \emph{Symbolic Data Analysis:
Conceptual Statistics and Data Mining}. Chichester, UK: John Wiley and Sons.

Billard, Lynne (2008). Sample covariance functions for complex quantitative
data. In \emph{Proceedings of the World IASC Conference}, pp. 157--163,
Yokohama, Japan.

Rodriguez-Rojas, Oldemar (2000). \emph{Classification et modeles lineaires
en analyse des donnees symboliques}. PhD thesis, Universite Paris IX
Dauphine.
}
\author{
Oldemar Rodriguez Rojas
}
\keyword{Covariance}
\keyword{Symbolic}
