% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInterval_boxplot.R
\name{ggInterval_boxplot}
\alias{ggInterval_boxplot}
\title{A interval Box plot}
\usage{
ggInterval_boxplot(data = NULL,mapping = aes(NULL),plotAll=FALSE,
width_type = "violin-like")
}
\arguments{
\item{data}{A ggInterval object. It can also be either RSDA object or
classical data frame, which will be automatically convert to ggInterval
data.}

\item{mapping}{Set of aesthetic mappings created by aes() or aes_().
If specified and inherit. aes = TRUE (the default),
it is combined with the default mapping at the top level of
the plot. You must supply mapping if there is no plot mapping.
It is the same as the mapping of ggplot2.}

\item{plotAll}{booleans, if TRUE, plot all variable together}

\item{width_type}{Box-width style. Use \code{"violin-like"} (default) to let
the rectangle widths reflect local interval concentration,
\code{"quantile-depth"} for progressively narrower quantile boxes, or
\code{"side-by-side"} for conventional boxplots of the lower and upper
bounds shown next to each other.}
}
\value{
Return a ggplot2 object.
}
\description{
Visualize the one continuous variable distribution using one of
three interval-aware boxplot styles.
}
\examples{
mydata <- ggInterval::facedata
ggInterval_boxplot(mydata, aes(x = AD))
ggInterval_boxplot(mydata, aes(x = AD), width_type = "quantile-depth")
ggInterval_boxplot(mydata, plotAll = TRUE, width_type = "side-by-side")

}
