% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInterval_CRplot.R
\name{ggInterval_CRplot}
\alias{ggInterval_CRplot}
\title{Center-range plot for interval data}
\usage{
ggInterval_CRplot(data = NULL,mapping = aes(NULL),plotAll=FALSE,
                         addEllipse = TRUE,
                         ellipseFill = "blue",
                         ellipseAlpha = 0.3)
}
\arguments{
\item{data}{A ggInterval object. It can also be either RSDA object or
classical data frame,which will be automatically convert to ggInterval
data.}

\item{mapping}{Set of aesthetic mappings created by aes() or aes_().
If specified and inherit. aes = TRUE (the default),
it is combined with the default mapping at the top level of
the plot. You must supply mapping if there is no plot mapping.
It is the same as the mapping of ggplot2.}

\item{plotAll}{booleans, if TRUE, plot all variable together}

\item{addEllipse}{logical, if TRUE (default), add a shaded ellipse layer.}

\item{ellipseFill}{fill color of the ellipse layer. Default is "blue".}

\item{ellipseAlpha}{alpha level of the ellipse layer. Default is 0.3.}
}
\value{
Return a ggplot2 object.
}
\description{
Visualize the relation between interval centers and ranges,
with mean reference lines and an optional ellipse overlay.
}
\examples{
mydata <- ggInterval::facedata
ggInterval_CRplot(mydata, aes(x = AD, col = "blue", shape = 2))
ggInterval_CRplot(mydata, plotAll = TRUE)
ggInterval_CRplot(mydata, plotAll = TRUE, addEllipse = FALSE)
}
