% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInterval_hist.R
\name{ggInterval_hist}
\alias{ggInterval_hist}
\title{Histogram for symbolic data with equal-bin or unequal-bin.}
\usage{
ggInterval_hist(data = NULL,mapping = aes(NULL),method="equal-bin",bins=10,
 plotAll = FALSE, position = "identity", alpha = 0.5)
}
\arguments{
\item{data}{A ggInterval object.It can also be either RSDA object or
classical data frame,which will be automatically convert to ggInterval
data.}

\item{mapping}{Set of aesthetic mappings created by aes() or aes_().
If specified and inherit. aes = TRUE (the default),
it is combined with the default mapping at the top level of
the plot. You must supply mapping if there is no plot mapping.
It is the same as the mapping of ggplot2.}

\item{method}{It can be \code{"equal-bin"} (default) or
\code{"unequal-bin"}. Equal-bin means the histogram uses equally spaced
intervals. Unequal-bin means the bin widths are determined by the data,
so the argument \code{bins} is unused in that case.}

\item{bins}{x axis bins,which mean how many partials the variable}

\item{plotAll}{boolean, whether plot all variables, default FALSE.
will be separate into.}

\item{position}{"stack" or "identity"}

\item{alpha}{fill alpha}
}
\value{
An object of class \code{ggInterval_hist_result}. A direct call
displays the histogram, and the returned object stores the ggplot object in
\code{$plot} together with the accompanying frequency tables.
}
\description{
Visualize the continuous variable distribution
by dividing the x axis into bins,and calculating the frequency
of observation interval in each bin.
}
\examples{
mydata <- ggInterval::facedata
ggInterval_hist(mydata, aes(x = AD), bins = 10)

hist_obj <- ggInterval_hist(mydata, plotAll = TRUE, bins = 10)
hist_obj
hist_obj$`Table AD`

ggInterval_hist(mydata, aes(x = AD), method = "unequal-bin")

}
