% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggInterval_MMplot.R
\name{ggInterval_MMplot}
\alias{ggInterval_MMplot}
\title{A min-max plot for interval data}
\usage{
ggInterval_MMplot(data = NULL,mapping = aes(NULL),
          scaleXY = "local",plotAll=FALSE)
}
\arguments{
\item{data}{A ggInterval object. It can also be either RSDA object or
classical data frame,which will be automatically convert to ggInterval
data.}

\item{mapping}{Set of aesthetic mappings created by aes() or aes_().
If specified and inherit. aes = TRUE (the default),
it is combined with the default mapping at the top level of
the plot. You must supply mapping if there is no plot mapping.}

\item{scaleXY}{default "local", which means limits of x-axis and y-axis
depend on their own variable. "global" means limits of them depend on all
variables that user input.}

\item{plotAll}{booleans, if TRUE, plot all variable together}
}
\value{
Return a ggplot2 object.
}
\description{
Visualize the range of the variables of each observations
by marking minimal and maximal point.
}
\examples{
mydata <- ggInterval::facedata
ggInterval_MMplot(mydata, aes(x = AD))
ggInterval_MMplot(mydata, aes(x = AD, size = 3))
ggInterval_MMplot(mydata, plotAll = TRUE, scaleXY = "global") +
  theme_classic()
}
