% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mushroom}
\alias{mushroom}
\title{mushroom data example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}, \code{symbolic_tbl}) with 23 rows and 3 columns.
}
\source{
Adapted from \code{RSDA::mushroom}; see
  \url{https://CRAN.R-project.org/package=RSDA}.
}
\usage{
data(mushroom)
}
\description{
mushroom interval data example.
}
\examples{
data(mushroom)
\donttest{
ggInterval_scatterplot(mushroom, aes(x = Cap.Widths, y = Stipe.Lengths))
}
}
\references{
Billard, Lynne and Diday, Edwin (2006).
\emph{Symbolic Data Analysis: Conceptual Statistics and Data Mining}.
Chichester, UK: John Wiley and Sons.
}
\keyword{datasets}
