% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oils}
\alias{oils}
\title{oils data example}
\format{
An object of class \code{symbolic_tbl} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8 rows and 4 columns.
}
\source{
Adapted from \code{RSDA::oils}; see
  \url{https://CRAN.R-project.org/package=RSDA}.
}
\usage{
data(oils)
}
\description{
oils interval data example.
}
\examples{
data(oils)
\donttest{
ggInterval_scatterplot(oils, aes(x = GRA, y = IOD))
}
}
\references{
Cazes, Pierre; Chouakria, Assia; Diday, Edwin; and
Schektman, Youri (1997). Extension de l'analyse en composantes principales
a des donnees de type intervalle. \emph{Revue de Statistique Appliquee},
45(3), 5--24.
}
\keyword{datasets}
