% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sd}
\alias{sd}
\alias{sd.default}
\alias{sd.symbolic_tbl}
\alias{sd.symbolic_interval}
\title{Generic function for the standard deviation}
\usage{
sd(x, ...)

\method{sd}{default}(x, na.rm = FALSE, ...)

\method{sd}{symbolic_tbl}(
  x,
  method = c("billard", "centers", "interval", "B", "BD", "BG"),
  na.rm = FALSE,
  ...
)

\method{sd}{symbolic_interval}(
  x,
  method = c("billard", "centers", "interval", "B", "BD", "BG"),
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{x}{First symbolic variables.}

\item{...}{As in R sd function.}

\item{na.rm}{As in R sd function.}

\item{method}{The method to use.}
}
\value{
Return a real number.
}
\description{
This function compute the symbolic standard deviation.
}
\details{
Supported interval-valued methods are:
\itemize{
  \item \code{"billard"}: standard deviation based on the Billard
  univariate variance formula.
  \item \code{"centers"}: standard deviation of interval centers.
  \item \code{"interval"}: interval-valued standard deviation obtained by
  standardizing lower and upper bounds separately.
  \item \code{"B"}: method-matched standard deviation defined by
  \eqn{\sqrt{C_B(X, X)}}.
  \item \code{"BD"}: method-matched standard deviation defined by
  \eqn{\sqrt{C_{BD}(X, X)}}.
  \item \code{"BG"}: method-matched standard deviation defined by
  \eqn{\sqrt{C_{BG}(X, X)}}.
}
}
\references{
Bertrand, Patrice and Goupil, Francoise (2000). Descriptive Statistics for
Symbolic Data. In Hans-Hermann Bock and Edwin Diday (eds.),
\emph{Analysis of Symbolic Data}, pp. 106--124. Berlin and Heidelberg:
Springer.

Billard, Lynne and Diday, Edwin (2006). \emph{Symbolic Data Analysis:
Conceptual Statistics and Data Mining}. Chichester, UK: John Wiley and Sons.

Billard, Lynne (2008). Sample covariance functions for complex quantitative
data. In \emph{Proceedings of the World IASC Conference}, pp. 157--163,
Yokohama, Japan.
}
\keyword{Symbolic}
\keyword{standard}
\keyword{deviation}
